requirements.txt
aiogram==3.3.0
asyncpg==0.29.0
python-dotenv==1.0.0
openai==1.3.5
anthropic==0.7.10
groq==0.3.0
aiohttp==3.9.1
python-dateutil==2.8.2
psycopg2-binary==2.9.9